refactor: webhooks package
